Compatibility changes

+++ Chess/ChessMove.h
@@ -28,9 +28,9 @@ public:
 	}
 
 	float vectorDistance() const {
-		return std::sqrtf(
-			std::powf(static_cast<float>(m_positionFrom.x() - m_positionTo.x()), 2) +
-			std::powf(static_cast<float>(m_positionFrom.y() - m_positionTo.y()), 2));
+		return std::sqrt(
+			std::pow(static_cast<float>(m_positionFrom.x() - m_positionTo.x()), 2) +
+			std::pow(static_cast<float>(m_positionFrom.y() - m_positionTo.y()), 2));
 	}
 
 	bool operator==(const ChessMove& rhs) const {
